+++ internal/app/subsystems/api/http/promise.go
@@ -158,7 +158,7 @@ type createPromiseAndTaskBody struct {
 
 type createPromiseTaskBody struct {
 	ProcessId string `json:"processId" binding:"required"`
-	Ttl       int    `json:"ttl" binding:"min=0"`
+	Ttl       int64  `json:"ttl" binding:"min=0"`
 }
 
 func (s *server) createPromiseAndTask(c *gin.Context) {
