+++ internal/app/subsystems/api/http/task.go
@@ -21,7 +21,7 @@ type claimTaskBody struct {
 	Id        string `json:"id" binding:"required"`
 	Counter   int    `json:"counter" binding:"required"`
 	ProcessId string `json:"processId" binding:"required"`
-	Ttl       int    `json:"ttl" binding:"min=0"`
+	Ttl       int64  `json:"ttl" binding:"min=0"`
 }
 
 func (s *server) claimTask(c *gin.Context) {
@@ -46,7 +46,7 @@ func (s *server) claimTask(c *gin.Context) {
 			Id:        c.Param("id"),
 			Counter:   counter,
 			ProcessId: s.api.TaskProcessId(c.Param("id"), counter),
-			Ttl:       int(s.config.TaskFrequency.Milliseconds()),
+			Ttl:       s.config.TaskFrequency.Milliseconds(),
 		}
 	} else {
 		var body claimTaskBody
